step3 순서도 이름변경 및 코드 네이밍 수정
